% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compositemh}
\alias{compositemh}
\title{This function refines the prior distributions of scaled heritability and the proportion of noncausal independent genes}
\usage{
compositemh(
  gent.data,
  ld.df,
  gent.Rho,
  gwasn,
  verbose = TRUE,
  niter = 10,
  thr_r2 = 0.5,
  min_size = 1,
  max_size = 100,
  max_K = 1e+06,
  max_r2 = 0.75,
  ...
)
}
\arguments{
\item{gent.data}{data frame of full gene-based association test results}

\item{ld.df}{data frame of weighted LD scores for all genes in gent.data}

\item{gent.Rho}{list of matrices. Each matrix in this list corresponds to a chromosome and represents correlations between gene-based association test statistics}

\item{gwasn}{GWAS sample size}

\item{verbose}{TRUE if progress should be printed, FALSE otherwise}

\item{niter}{number of iterations for which random sampling of independent genes should be done}

\item{thr_r2}{genes correlated beyond this threshold will be placed in the same block; the thr_r2 argument in bigsnpr::snp_ldsplit}

\item{min_size}{the minimum size of blocks of genes; the min_size argument in bigsnpr::snp_ldsplit}

\item{max_size}{the maximum size of blocks of genes; the min_size argument in bigsnpr::snp_ldsplit}

\item{max_K}{the maximum number of gene blocks; the max_K argument in bigsnpr::snp_ldsplit}

\item{max_r2}{the maximum allowable correlation between genes in separate blocks; the max_r2 argument in bigsnpr::snp_ldsplit}

\item{...}{additional arguments passed to the mh function for prior estimation}
}
\description{
This function refines the prior distributions of scaled heritability and the proportion of noncausal independent genes
}
\examples{
compositemh()
}
\keyword{prior}
