% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dtrap}
\alias{dtrap}
\title{Trapezoidal density}
\usage{
dtrap(x, a, x1, x2, b, log = FALSE)
}
\arguments{
\item{x}{argument to the density function}

\item{a}{smallest parameter of the trapezoidal density function}

\item{x1}{second-smallest parameter of the trapezoidal density function}

\item{x2}{second-largest parameter of the trapezoidal density function}

\item{b}{largest parameter of the trapezoidal density function}

\item{log}{TRUE if log density should be returned, FALSE otherwise}
}
\description{
This function calculates the density of the trapezoidal distribution
}
\examples{
dtrap(1,0,1,2,3)
dtrap()
}
\keyword{trapezoidal}
