% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mom_beta_distribution}
\alias{mom_beta_distribution}
\title{This function uses the method of moments to estimate the parameters of a Beta distribution from observed values}
\usage{
mom_beta_distribution(deltas, var_deltas = NULL)
}
\arguments{
\item{deltas}{estimate of the variance of the observed values}
}
\description{
This function uses the method of moments to estimate the parameters of a Beta distribution from observed values
}
\examples{
mom_beta_distribution(c(1,3,4,5,7)/10,1/20)
mom_beta_distribution()
}
\keyword{Beta}
\keyword{MoM}
