% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mom_trap_distribution}
\alias{mom_trap_distribution}
\title{This function uses the method of moments approximation to estimate the parameters of a trapezoidal distribution from observed values when the boundaries of the observable space are known}
\usage{
mom_trap_distribution(taus, a, b)
}
\arguments{
\item{taus}{vector of observed random variables following a trapezoidal distribution}

\item{a}{smallest parameter of the trapezoidal density function}

\item{b}{largest parameter of the trapezoidal density function}
}
\description{
This function uses the method of moments approximation to estimate the parameters of a trapezoidal distribution from observed values when the boundaries of the observable space are known
}
\examples{
mom_trap_distribution(c(1,3,4,5,7),0,8)
numerical_trap_distribution()
}
\keyword{MoM}
\keyword{trapezoidal}
