% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{posterior_gene}
\alias{posterior_gene}
\title{This function returns gene-level posterior probabilities of causality for a single trait that do not consider correlation between genes}
\usage{
posterior_gene(gent.data, ld.df, gwasn, prior.estimation.list, verbose = TRUE)
}
\arguments{
\item{gent.data}{data frame of full gene-based association test results}

\item{ld.df}{data frame of weighted LD scores for all genes in gent.data}

\item{gwasn}{GWAS sample size}

\item{prior.estimation.list}{direct output of the compositemh function containing information regarding the prior distributions to use in posterior estimation}

\item{verbose}{TRUE if progress should be printed, FALSE otherwise}

\item{gent.Rho}{list of matrices. Each matrix in this list corresponds to a chromosome and represents correlations between gene-based association test statistics}
}
\description{
This function returns gene-level posterior probabilities of causality for a single trait that do not consider correlation between genes
}
\examples{
posterior_gene()
}
\keyword{posterior}
