% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{propshared}
\alias{propshared}
\title{This function estimates the counts and proportions of independent causal genes which are shared between two traits}
\usage{
propshared(
  posteriors1,
  posteriors2,
  gwasn1,
  gwasn2,
  gent.Rho,
  niter = 10,
  verbose = TRUE,
  thr_r2 = 0.5,
  min_size = 1,
  max_size = 30,
  max_K = 1e+06,
  max_r2 = 0.75
)
}
\arguments{
\item{posteriors1}{direct output of posterior_gene applied to trait 1}

\item{posteriors2}{direct output of posterior_gene applied to trait 2}

\item{gwasn1}{GWAS sample size of trait 1}

\item{gwasn2}{GWAS sample size of trait 2}

\item{gent.Rho}{list of matrices. Each matrix in this list corresponds to a chromosome and represents correlations between gene-based association test statistics}

\item{niter}{number of iterations for which random sampling of independent genes should be done}

\item{verbose}{TRUE if progress should be printed, FALSE otherwise}

\item{thr_r2}{genes correlated beyond this threshold will be placed in the same block; the thr_r2 argument in bigsnpr::snp_ldsplit}

\item{min_size}{the minimum size of blocks of genes; the min_size argument in bigsnpr::snp_ldsplit}

\item{max_size}{the maximum size of blocks of genes; the min_size argument in bigsnpr::snp_ldsplit}

\item{max_K}{the maximum number of gene blocks; the max_K argument in bigsnpr::snp_ldsplit}

\item{max_r2}{the maximum allowable correlation between genes in separate blocks; the max_r2 argument in bigsnpr::snp_ldsplit}
}
\description{
This function estimates the counts and proportions of independent causal genes which are shared between two traits
}
\examples{
propshared()
}
\keyword{shared}
