% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qtrap}
\alias{qtrap}
\title{This function numerically calculates the quantile function of the trapezoidal distribution}
\usage{
qtrap(p, a, x1, x2, b, eps = 1e-05, max_passes = 5, n = 100)
}
\arguments{
\item{p}{the desired quantile defines this level of probability}

\item{a}{smallest parameter of the trapezoidal density function}

\item{x1}{second-smallest parameter of the trapezoidal density function}

\item{x2}{second-largest parameter of the trapezoidal density function}

\item{b}{largest parameter of the trapezoidal density function}

\item{eps}{error in numerical approximation of the true quantile}

\item{max_passes}{number of intervals, of decreasing width, in which the quantile should be approximated}

\item{n}{number of grid points at which to evaluate the quantile in each interval}
}
\description{
This function numerically calculates the quantile function of the trapezoidal distribution
}
\examples{
qtrap(0.5,0,1,2,3)
qtrap()
}
\keyword{trapezoidal}
