% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rtrap}
\alias{rtrap}
\title{This function geometrically samples from the trapezoidal distribution}
\usage{
rtrap(n, a, x1, x2, b)
}
\arguments{
\item{n}{the desired number of random draws}

\item{a}{smallest parameter of the trapezoidal density function}

\item{x1}{second-smallest parameter of the trapezoidal density function}

\item{x2}{second-largest parameter of the trapezoidal density function}

\item{b}{largest parameter of the trapezoidal density function}
}
\description{
This function geometrically samples from the trapezoidal distribution
}
\examples{
rtrap(10,0,1,2,3)
rtrap()
}
\keyword{trapezoidal}
