% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shared_count_simex}
\alias{shared_count_simex}
\title{This function adjusts the raw estimates of shared causal genes between a pair of traits for GWAS sample overlap}
\usage{
shared_count_simex(
  M,
  mcausalgenes1,
  mcausalgenes2,
  ngwas1,
  ngwas2,
  estimated_overlap_count,
  upsilon_overlap,
  h21 = 0.2,
  h22 = 0.2,
  niter = 1000,
  m = 50,
  mcausal_snps_per_gene = 3,
  LD_type = "AR",
  LD_rho = 0.5,
  nlambdas = 10,
  doplot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{M}{number of genes tested for association with each trait}

\item{mcausalgenes1}{estimated number of genes causing trait 1}

\item{mcausalgenes2}{estimated number of genes causing trait 2}

\item{ngwas1}{sample size of trait 1 GWAS}

\item{ngwas2}{sample size of trait 2 GWAS}

\item{estimated_overlap_count}{raw data-estimated number of overlapping causal genes}

\item{upsilon_overlap}{sample overlap correlation coefficient, ie correlation between SNP effect sizes due to sample overlap. See LeBlanc et al. 2015}

\item{h21}{estimated/assumed SNP heritability of trait 1}

\item{h22}{estimated/assumed SNP heritability of trait 2}

\item{niter}{number of iterations in each simulation}

\item{m}{number of tested SNPs per gene}

\item{mcausal_snps_per_gene}{assumed number of causal SNPs per causal gene}

\item{LD_type}{structure of correlations between gene-based test statistics due to shared LD of their SNP sets. AR for first order autoregressive, CS for compound symmetry, or I for independence.}

\item{LD_rho}{correlation parameter specifying the matrix of correlations between gene-based test statistics}

\item{nlambdas}{number of simulated sample overlap correlation values to evaluate in SIMEX}

\item{doplot}{TRUE if a plot should be generated, FALSE otherwise}

\item{verbose}{TRUE if running results should be printed, FALSE otherwise}
}
\description{
This function adjusts the raw estimates of shared causal genes between a pair of traits for GWAS sample overlap
}
\examples{
shared_count_simex()
}
\keyword{SIMEX}
